# docs/swagger.yaml
basePath: /api
definitions:
...
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.CursorPagination:
    properties:
      next:
        example: cba
        type: string
      prev:
        example: abc
        type: string
      totalEntries:
        example: 100
        type: integer
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel:
    properties:
      code: {}
      message:
        example: error
        type: string
      status:
        example: error
        type: string
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel:
    properties:
      errors: {}
      message:
        example: validation error
        type: string
      status:
        example: error
        type: string
    type: object
  ? bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestPaginationResponseModel-array_bitbucket_org_Amartha_go-fp-transaction_internal_models_DoGetTransactionResponse
  : properties:
      contents:
        items:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.DoGetTransactionResponse'
        type: array
      kind:
        example: collection
        type: string
      pagination:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.CursorPagination'
    type: object
...
  ? bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestPaginationResponseModel-array_bitbucket_org_Amartha_go-fp-transaction_internal_models_ListWalletTrxByAccountNumberResponse
  : properties:
      contents:
        items:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.ListWalletTrxByAccountNumberResponse'
        type: array
      kind:
        example: collection
        type: string
      pagination:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.CursorPagination'
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_common_validation.ErrorValidateResponse:
    properties:
      code:
        example: accountNumber_required
        type: string
      field:
        example: MISSING_FIELD
        type: string
      message:
        example: field is missing
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.Amount:
    properties:
      currency:
        type: string
      value:
        type: number
    required:
    - value
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.AmountDetail:
    properties:
      amount:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.Amount'
      type:
        type: string
    required:
    - amount
    - type
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.CreateWalletTransactionRequest:
    properties:
      accountNumber:
        type: string
      amounts:
        items:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.AmountDetail'
        type: array
      clientId:
        description: internal use
        type: string
      description:
        type: string
      destinationAccountNumber:
        type: string
      idempotencyKey:
        type: string
      isReserved:
        type: boolean
      metadata:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletMetadata'
      netAmount:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.Amount'
      refNumber:
        type: string
      transactionFlow:
        allOf:
        - $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionFlow'
        enum:
        - cashin
        - cashout
        - transfer
        - refund
      transactionTime:
        type: string
      transactionType:
        type: string
    required:
    - accountNumber
    - netAmount
    - refNumber
    - transactionFlow
    - transactionTime
    - transactionType
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.DoCreateTransactionRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      fromAccount:
        type: string
      isReserved:
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      method:
        type: string
      orderType:
        type: string
      refNumber:
        type: string
      toAccount:
        type: string
      transactionTime:
        type: string
      transactionType:
        type: string
    required:
    - amount
    - fromAccount
    - orderType
    - refNumber
    - toAccount
    - transactionTime
    - transactionType
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.DoGetTransactionResponse:
    properties:
      amount:
        example: "50777"
        type: string
      createdAt:
        example: "2006-01-02 15:04:05"
        type: string
      currency:
        example: IDR
        type: string
      description:
        example: Topup from VA
        type: string
      fromAccount:
        example: "189513"
        type: string
      fromAccountName:
        example: John
        type: string
      fromAccountProductTypeName:
        example: valid
        type: string
      kind:
        example: transaction
        type: string
      metadata:
        additionalProperties:
          type: string
        example:
          key: value
        type: object
      method:
        example: TOPUP.VA
        type: string
      orderType:
        example: TOPUP
        type: string
      orderTypeName:
        example: TOPUP
        type: string
      refNumber:
        example: 55aa66bb-e6e0-4065-9f4a-64182e97e9d9
        type: string
      status:
        example: "1"
        type: string
      toAccount:
        example: "222000000069"
        type: string
      toAccountName:
        example: John
        type: string
      toAccountProductTypeName:
        example: valid
        type: string
      transactionDate:
        example: "2023-10-25"
        type: string
      transactionId:
        example: c172ca84-9ae2-489c-ae4f-8ef372a109ae
        type: string
      transactionTime:
        example: "2023-10-25 08:08:26"
        type: string
      transactionType:
        example: TOPUP
        type: string
      transactionTypeName:
        example: TOPUP
        type: string
      updatedAt:
        example: "2006-01-02 15:04:05"
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.ListWalletTrxByAccountNumberResponse:
    properties:
      amount:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.Amount'
      amounts:
        items:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.AmountDetail'
        type: array
      description:
        type: string
      kind:
        type: string
      metadata:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletMetadata'
      refNumber:
        type: string
      status:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletTransactionStatus'
      transactionDate:
        type: string
      transactionFlow:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionFlow'
      transactionTime:
        type: string
      transactionType:
        type: string
      transactionWalletId:
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionFlow:
    enum:
    - cashin
    - cashout
    - transfer
    - refund
    type: string
    x-enum-varnames:
    - TransactionFlowCashIn
    - TransactionFlowCashOut
    - TransactionFlowTransfer
    - TransactionFlowRefund
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusReservedTransactionRequest:
    properties:
      status:
        enum:
        - commit
        - cancel
        example: commit
        type: string
    required:
    - status
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusReservedTransactionResponse:
    properties:
      kind:
        example: transaction
        type: string
      status:
        example: SUCCESS
        type: string
      transactionId:
        example: 41d03147-c017-4176-8a1a-0b7ec735cc29
        type: string
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusWalletTransactionRequest:
    properties:
      action:
        enum:
        - commit
        - cancel
        example: commit
        type: string
      clientId:
        description: internal use
        type: string
      transactionTime:
        type: string
    required:
    - action
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusWalletTransactionResponse:
    properties:
      kind:
        example: walletTransaction
        type: string
      status:
        example: SUCCESS
        type: string
      transactionId:
        example: 41d03147-c017-4176-8a1a-0b7ec735cc29
        type: string
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletMetadata:
    additionalProperties: {}
    type: object
...
  bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletTransactionResponse:
    properties:
      accountNumber:
        type: string
      amounts:
        items:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.AmountDetail'
        type: array
      description:
        type: string
      destinationAccountNumber:
        type: string
      id:
        type: string
      kind:
        type: string
      metadata:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletMetadata'
      netAmount:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.Amount'
      refNumber:
        type: string
      status:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletTransactionStatus'
      transactionFlow:
        $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.TransactionFlow'
      transactionTime:
        type: string
      transactionType:
        type: string
    type: object
  bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletTransactionStatus:
    enum:
    - PENDING
    - SUCCESS
    - CANCEL
    type: string
    x-enum-varnames:
    - WalletTransactionStatusPending
    - WalletTransactionStatusSuccess
    - WalletTransactionStatusCancel
...
host: localhost:9567
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a go fp transaction api docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GO FP TRANSACTION API DUCUMENTATION
  version: "1.0"
paths:
...
  /transaction:
    post:
      consumes:
      - application/json
      description: Create data from any transaction
      parameters:
      - description: A JSON object containing create transaction payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.DoCreateTransactionRequest'
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.DoGetTransactionResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            create transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "404":
          description: Data not found. This can happen if there is a data not found
            while create transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "422":
          description: Validation error. This can happen if there is an error validation
            while create transaction
          schema:
            allOf:
            - $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_validation.ErrorValidateResponse'
                  type: array
              type: object
        "500":
          description: Internal server error. This can happen if there is an error
            while create transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Create data transaction
      tags:
      - Transaction
...
  /v1/internal-wallets/accounts/{accountNumber}/transactions:
    get:
      consumes:
      - application/json
      description: Get all wallet transaction by accountNumber
      parameters:
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - example: "21100100000001"
        in: query
        name: accountNumber
        required: true
        type: string
      - example: "2023-01-07"
        in: query
        name: endDate
        type: string
      - example: 10
        in: query
        name: limit
        type: integer
      - example: abc
        in: query
        name: nextCursor
        type: string
      - example: cba
        in: query
        name: prevCursor
        type: string
      - in: query
        name: sortBy
        type: string
      - in: query
        name: sortDirection
        type: string
      - example: "2023-01-01"
        in: query
        name: startDate
        type: string
      - enum:
        - CANCEL
        - PENDING
        - SUCCESS
        example: PENDING
        in: query
        name: status
        type: string
      - example: TUPVA
        in: query
        name: transactionType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestPaginationResponseModel-array_bitbucket_org_Amartha_go-fp-transaction_internal_models_ListWalletTrxByAccountNumberResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get all wallet transaction by accountNumber
      tags:
      - Internal Wallet
  /v1/internal-wallets/accounts/transactions:
    get:
      consumes:
      - application/json
      description: Get all wallet transaction
      parameters:
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - example: 21100100000001,21100100000002,21100100000003
        in: query
        name: accountNumbers
        type: string
      - example: "2023-01-07"
        in: query
        name: endDate
        type: string
      - example: 10
        in: query
        name: limit
        type: integer
      - example: abc
        in: query
        name: nextCursor
        type: string
      - example: cba
        in: query
        name: prevCursor
        type: string
      - in: query
        name: sortBy
        type: string
      - in: query
        name: sortDirection
        type: string
      - example: "2023-01-01"
        in: query
        name: startDate
        type: string
      - enum:
        - CANCEL
        - PENDING
        - SUCCESS
        example: PENDING
        in: query
        name: status
        type: string
      - example: TUPVA
        in: query
        name: transactionType
        type: string
      - example: VOLTR,TUPVA
        in: query
        name: transactionTypes
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestPaginationResponseModel-array_bitbucket_org_Amartha_go-fp-transaction_internal_models_ListWalletTrxByAccountNumberResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get all wallet transaction
      tags:
      - Internal Wallet
...
  /v1/transaction:
    get:
      consumes:
      - application/json
      description: Get All transaction
      parameters:
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - example: "2023-01-07"
        in: query
        name: endDate
        type: string
      - example: 10
        in: query
        name: limit
        type: integer
      - example: abc
        in: query
        name: nextCursor
        type: string
      - example: TOPUP
        in: query
        name: orderType
        type: string
      - example: cba
        in: query
        name: prevCursor
        type: string
      - example: Poket
        in: query
        name: productTypeName
        type: string
      - example: value of accountNumber refNumber transactionId
        in: query
        name: search
        type: string
      - example: accountNumber refNumber transactionId
        in: query
        name: searchBy
        type: string
      - example: "2023-01-01"
        in: query
        name: startDate
        type: string
      - description: 'Deprecated: /v1/transaction not use this anymore'
        example: TOPUP
        in: query
        name: transactionType
        type: string
      - collectionFormat: csv
        example:
        - TOPUP
        in: query
        items:
          type: string
        name: transactionTypes
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestPaginationResponseModel-array_bitbucket_org_Amartha_go-fp-transaction_internal_models_DoGetTransactionResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get All transaction
      tags:
      - Transactions
...
  /v1/transaction/{transactionId}:
    patch:
      consumes:
      - application/json
      description: Update status reserved transaction
      parameters:
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - description: transactionId
        in: path
        name: transactionId
        required: true
        type: string
      - description: A JSON object containing payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusReservedTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusReservedTransactionResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Update status reserved transaction
      tags:
      - Transaction
  /v1/transaction/{transactionType}/{refNumber}:
    get:
      consumes:
      - application/json
      description: Get single transaction with criteria
      parameters:
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      - description: transactionType
        in: path
        name: transactionType
        required: true
        type: string
      - description: refNumber
        in: path
        name: refNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.DoGetTransactionResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "500":
          description: Internal server error. This can happen if there is an error
            while get all transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Get single transaction with criteria
      tags:
      - Transactions
...
  /wallet-transactions:
    post:
      consumes:
      - application/json
      description: Create wallet transaction
      parameters:
      - description: A JSON object containing create transaction payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.CreateWalletTransactionRequest'
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.WalletTransactionResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            create transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "404":
          description: Data not found. This can happen if there is a data not found
            while create transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "422":
          description: Validation error. This can happen if there is an error validation
            while create transaction
          schema:
            allOf:
            - $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_validation.ErrorValidateResponse'
                  type: array
              type: object
        "500":
          description: Internal server error. This can happen if there is an error
            while create transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Create wallet transaction
      tags:
      - WalletTransaction
  /wallet-transactions/{id}:
    patch:
      consumes:
      - application/json
      description: Update status wallet transaction
      parameters:
      - description: A JSON object containing create transaction payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusWalletTransactionRequest'
      - description: X-Secret-Key
        in: header
        name: X-Secret-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response indicates that the request succeeded and the resources
            has been fetched and transmitted in the message body
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_models.UpdateStatusWalletTransactionResponse'
        "400":
          description: Bad request error. This can happen if there is an error while
            update status transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "404":
          description: Data not found. This can happen if there is a data not found
            while update status transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
        "422":
          description: Validation error. This can happen if there is an error validation
            while update status transaction
          schema:
            allOf:
            - $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorValidationResponseModel'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_validation.ErrorValidateResponse'
                  type: array
              type: object
        "500":
          description: Internal server error. This can happen if there is an error
            while update status transaction
          schema:
            $ref: '#/definitions/bitbucket_org_Amartha_go-fp-transaction_internal_common_http.RestErrorResponseModel'
      summary: Update status wallet transaction
      tags:
      - WalletTransaction
schemes:
- http
swagger: "2.0"